refactor(flowmux): extract UDP wire helpers
